refactor(Message): 코드 리뷰 반영
- TextEditor에서 폰트 반영 관련해서 useEffect와 인라인 중복 코드 발생해서 인라인 코드 제거
- Router에 MessagePage 반영하여 import Header 제거
