+++ src/Router.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import EditPage from "./pages/EditPage";
 import HomePage from "./pages/HomePage/HomePage";
 import ListPage from "./pages/ListPage/ListPage";
-import MessagePage from "./pages/MessagePage";
+import MessagePage from "./pages/MessagePage/MessagePage";
 import NotFoundPage from "./pages/NotFoundPage";
 import PostItemPage from "./pages/PostItemPage";
 import PostPage from "./pages/PostPage";
